% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroecologyRexercise.R
\docType{package}
\name{macroecologyRexercise}
\alias{macroecologyRexercise}
\alias{macroecologyRexercise-package}
\title{macroecologyRexercise: A package of datasets for the macroecology R exercises.}
\description{
macroecologyRexercise: A package of datasets for the macroecology R exercises.
}
