% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{palms}
\alias{palms}
\title{Palm Species Occurrences in the World Botanical Regions}
\format{An object of class `data.frame`.}
\source{
\link{https://www.nature.com/articles/s41597-019-0189-0}
}
\usage{
palms
}
\description{
A `data.frame` for given the name of botanical regions and the palm species name.
}
\details{
See \link{https://www.nature.com/articles/s41597-019-0189-0} for details.
}
\examples{
data(palms)
plot(palms)
}
\keyword{datasets}
