% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{phylacine}
\alias{phylacine}
\title{Data from Phylacine: The Phylogenetic Atlas of Mammal Macroecology}
\format{An object of class `data.frame`.}
\source{
\link{https://megapast2future.github.io/PHYLACINE_1.2/}
}
\usage{
phylacine
}
\description{
A `data.frame` with information on species traits (body mass, diet, life habit) and more for all extant and recently extinct mammal species.
}
\details{
See \link{https://megapast2future.github.io/PHYLACINE_1.2/} for details.
}
\examples{
data(phylacine)
}
\keyword{datasets}
