% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tdwg}
\alias{tdwg}
\title{World Botanical Regions}
\format{An object of class `SpatialPolygonsDataFrame`.}
\source{
\link{https://www.nature.com/articles/s41597-019-0189-0}
}
\usage{
tdwg
}
\description{
A `SpatialPolygonsDataFrame` for world botanical regions.
}
\details{
See \link{https://www.nature.com/articles/s41597-019-0189-0} for details.
}
\examples{
data(tdwg)
plot(tdwg)
}
\keyword{datasets}
